Move cache traits
