+++ crates/relayer-framework/src/base/builder/traits/mod.rs
@@ -1,4 +1,5 @@
 pub mod birelay;
 pub mod builder;
+pub mod cache;
 pub mod chain;
 pub mod relay;
