+++ crates/relayer-framework/src/base/one_for_all/impls/builder/mod.rs
@@ -1,4 +1,5 @@
 pub mod birelay;
+pub mod cache;
 pub mod chain;
 pub mod error;
 pub mod relay;
